# go/model_article.go
/*
 * Article
 *
 * This is a simple API Server based DDD
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

import (
	"time"
)

type Article struct {
	Id int64 `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	Content string `json:"content,omitempty"`

	Username string `json:"username,omitempty"`

	CreatedAt time.Time `json:"createdAt,omitempty"`

	UpdatedAt time.Time `json:"updatedAt,omitempty"`
}